last fixes

+++ src/transformers/models/eomt_dinov3/modeling_eomt_dinov3.py
@@ -1257,7 +1257,7 @@ class EomtDinov3ForUniversalSegmentation(EomtDinov3PreTrainedModel):
         masks_queries_logits_per_layer, class_queries_logits_per_layer = (), ()
 
         hidden_states = self.dropout(self.embeddings(pixel_values))
-        position_embeddings = self.rope_embeddings(pixel_values)
+        position_embeddings = self.rope_embeddings(pixel_values.to(hidden_states.dtype))
         attention_mask = None
 
         for idx, layer_module in enumerate(self.layers):
